chore: switch back to pages for html support
