+++ log-app/package.json
@@ -1,18 +1,13 @@
 {
-  "name": "log-app",
+  "name": "cf-pages-functions",
   "version": "0.0.0",
-  "devDependencies": {
-    "@cloudflare/workers-types": "^4.20230814.0",
-    "typescript": "^5.1.6",
-    "wrangler": "3.4.0"
-  },
-  "private": true,
   "scripts": {
-    "start": "wrangler dev",
-    "deploy": "wrangler deploy"
+    "build": "tsc"
   },
-  "dependencies": {
-    "@types/ws": "^8.5.5",
-    "ws": "^8.13.0"
-  }
+  "devDependencies": {
+    "@cloudflare/workers-types": "^3.2.0",
+    "@supabase/supabase-js": "^2.32.0",
+    "typescript": "^4.3.2"
+  },
+  "dependencies": {}
 }
