+++ log-app/types/global.ts
@@ -1,0 +1,5 @@
+export interface Env {
+  ASSETS: Fetcher;
+  SUPABASE_KEY: string;
+  SUPABASE_URL: string;
+}
